move dex prices to api
